pyproject.toml
[project]
name = "games-report-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcos Paulo",email = "mpaulosp.315@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standart] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ijson (>=3.3.0,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
